rename: features 폴더 정리
