+++ src/entities/shared-game/ui/sort-tab/sort-tabs.stories.ts
