+++ src/entities/shared-game/ui/sort-tab/sort-tabs.tsx
